refactor(opinion thunk tests): moved them out of user spec and into their own file
